TransportDelegate identifierUpdated reports whether identifier is host
allows us to distinguish the client from the host role.   Generally,
this is necessary to allow us to deterministically choose one peer in a
two peer connection as responsible for initiating the flow.

This wasn’t necessary for BLEBluetoothTransport because previously
BLEPeripheral did not report connection events. This now occurs via
CBPeripheralManager didSubscribeToCharacteristic.